Added new Simple Mailer example and updated go-mail version
A new example code for a Simple Mailer was added to `content/en/examples/simple-mailer.md` for better understanding of the go-mail usage. The example was also added to the main menu. Additionally, the version of `go-mail` used was updated from v0.3.1 to v0.4.1 to keep up with the latest changes and improvements.